refactor: move webhook server execution in a dedicated package
